PMA-4 fix(header): fix edit profile form
