--- src/components/App/styles.module.scss
@@ -1,9 +1,0 @@
-.testClass {
-  color: #ff0000;
-}
-
-/* .wrapper {
-  width: calc(100% - 100px);
-  margin: 0 auto;
-}
- */
